Seed db by scraping github issues
Scrapes github for open issues tagged with the 'good first issue' label.

There is a problem in that github limits searches to 1000 results. Therefore,
the scraper only returns the most recent 1000 issues for each language.

Perhaps this would be better split into two stages: write the scraped data to
a file, then seed the db with that written data.