Add yaml file of all known languages on Github
This will be used to get the colours associated with each language on Github.
These will be used as the background colours of the language tags in the list
view.

Note: The scraper already fetches a list of languages from Github separately but I don't yet know how to fetch a list of colours through the api.

The source of this yaml file is: https://raw.githubusercontent.com/github/linguist/master/lib/linguist/languages.yml