Add space between pagination links and issues
I've applied the styling as two separate classes. There may be a cleaner way
of doing it. I tried using pseudo selectors to select the first and last
occurrences of `.pagination` but it didn't seem to work.

Moving on because I don't want to focus too much on the styling at the moment.