Order issues by creation date
It makes sense to me (at the moment) to order the posts by date created. At
the moment, this date in the db matches the creation date of the issue on
Github.

I could also record the date of the last update in the db… but I have not yet
done that. Maybe later.