Allow issue to be looked up by intersection of language and labels
I've introduced a new scope with this. I had hoped to be able to perform the
query just by `by_language(languge).or by_issues(issue)` but the two relations
return by the scopes are structurally incompatible.
https://api.rubyonrails.org/classes/ActiveRecord/QueryMethods.html#method-i-or

I may yet move language to its own table but I'm committing now because I've
spent a lot of time on this and want to move forwards.