Ensure all labels are returned for each issue row
I can search for an issue by a single label like so:

```ruby
> Issue.by_labels('bug').first.labels.map(&:name)
=> ["bug", "help wanted"]
```

In the controller I want to eager load labels so not to end up with N+1
queries. I use `includes`:

```ruby
@issues = Issue.by_labels('bug').includes(:labels, :language)
```

However, I now see that only the label specified in the where query is
returned:

```ruby
> Issue.by_labels('bug').includes(:labels, :language)
                        .first.labels.map(&:name)
=> ["bug"]
```

I need each issue row to point to all of its labels. `["bug", "help wanted"]`
not `["bug"]`.

I don't quite know why it works but passing a string to `includes` instead of
a symbol returns both labels.

```ruby
> Issue.by_labels('bug').includes('labels', 'language')
                        .first.labels.map(&:name)
=> ["bug", "help wanted"]
```

I wrote a regression test but could not get it to fail. It passed no matter
whether `includes` was supplied with strings on symbols. I have no idea why
and so have commented the test out as it currently useless. Instead, I have
tested for regression in the view. Since this test fails when `includes` is
passed symbols, I am happy to continue.