Link to issues in tags view
In a forthcoming commit I will change the behaviour of the tags in the tags
view. Currently, clicking a tag navigates the user to the issues page. I will
change this so that the tag is edited to the selected list. The user will have
to explicitly navigate to the issues page.