AJAXify label filtering
The filtered labels are rendered as a result of the filter form submission.
The params need to be sent with the form.

Each time a label is added, the form needs to be reloaded so that the params
can be set in the target URL. Then, when the tags are loaded via AJAX, they
too will have the params in their URLs.

(And other stuff probably but I'm too tired to break it into separate
(commits.)