Refactor application helper as step towards using session storage
Passing around the params as a way of persisting state is getting out of hand.
That is, having to make sure the params are sent with every request is
becoming a little fragile. Any link that doesn't, wipes the slate clean. I
think using session storage would be a logical next step.