Store selected tags in session hash
Messy code here. I'm running out of time and I need to move onto something else.
Tags cannot be removed unless an empty string `['']` is sent in the params
`labels: ['']`. Otherwise the controller will look at the params and see `nil`
-- for instance, the pagination links don't send labels or language params --
and decide not to change the session values.

It leaves the code more fragile but…

I've probably been thinking about this the wrong way round. I went with params early on because of the simplicity but I've diverged from doing things the rails way. With :language and :labels I've got some notion of state but there's no model to represent them. I'd avoided that becuase I didn't want to burden the database but, come to think of it, I don't need a table to back the model.

That's the next step then. If I introduce a new model, at least I can hide this mess in one place.